Move to trait, to reuse downstream
