+++ layers/GraphQLByPoP/packages/graphql-clients-for-wp/src/ConditionalOnContext/UseGraphiQLExplorer/Overrides/Services/Clients/GraphiQLWithExplorerClient.php
@@ -8,22 +8,5 @@ use GraphQLByPoP\GraphQLClientsForWP\Clients\AbstractGraphiQLClient;
 
 class GraphiQLWithExplorerClient extends AbstractGraphiQLClient
 {
-    protected function getClientRelativePath(): string
-    {
-        // This is where the files are generated when running `npm run build`
-        return '/clients/graphiql-with-explorer/build';
-    }
-    protected function getJSFilename(): string
-    {
-        return 'main.js';
-    }
-    /**
-     * Assets folder name
-     */
-    protected function getAssetDirname(): string
-    {
-        // Please notice! GraphiQL Explorer loads under "/assets...",
-        // that's why the dirname starts with "/"
-        return '/static';
-    }
+    use GraphiQLWithExplorerClientTrait;
 }
